fix : 로직 최적화
